+++ design-system/checkbox/src/internal/label.tsx
@@ -98,6 +98,7 @@ export default function Label({
   const { mode } = useGlobalTheme();
 
   return (
+    // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-noninteractive-element-interactions
     <label
       css={[
         labelStyles,