+++ editor/editor-wikimarkup-transformer/src/parser/tokenize/keyword.ts
@@ -6,7 +6,7 @@ interface KeywordToken {
   regex: RegExp;
 }
 
-const macroKeywordTokenMap: KeywordToken[] = [
+export const macroKeywordTokenMap: KeywordToken[] = [
   {
     type: TokenType.ADF_MACRO,
     regex: /^{adf/i,